+++ prettier.config.mjs
@@ -1,0 +1,3 @@
+import { createConfig } from '@stancer/prettier';
+
+export default createConfig();