+++ scripts/messages/compile.js
@@ -1,12 +1,8 @@
-const exec = require("node:child_process").exec;
-const path = require("node:path");
-const glob = require("glob");
+const exec = require('node:child_process').exec;
+const path = require('node:path');
+const { glob } = require('glob');
 
-glob("languages/*.po", (err, files) => {
-  if (err) {
-    throw err;
-  }
-
+glob('languages/*.po').then((files) => {
   files.forEach((file) => {
     const portable = path.join(process.cwd(), file);
     const message = portable.replace(/\.po$/, '.mo');