+++ scripts/messages/merge.js
@@ -1,13 +1,9 @@
-const fs = require("node:fs");
-const exec = require("node:child_process").exec;
-const path = require("node:path");
-const glob = require("glob");
+const fs = require('node:fs');
+const exec = require('node:child_process').exec;
+const path = require('node:path');
+const { glob } = require('glob');
 
-glob("languages/*.po", (err, files) => {
-  if (err) {
-    throw err;
-  }
-
+glob('languages/*.po').then((files) => {
   files.forEach((file) => {
     const portable = path.join(process.cwd(), file);
     const template = portable.replace(/-[a-z]{2}_[A-Z]{2}\.po$/, '.pot');
@@ -18,10 +14,9 @@ glob("languages/*.po", (err, files) => {
       }
 
       const current = new Date().getFullYear();
-      const content = fs.readFileSync(portable, { encoding: 'utf-8' })
-        .replace(/^# Copyright \(C\) .*/m, `# Copyright (C) 2023-${current} Stancer / Iliad 78`)
-      ;
-
+      const content = fs
+        .readFileSync(portable, { encoding: 'utf-8' })
+        .replace(/^# Copyright \(C\) .*/m, `# Copyright (C) 2023-${current} Stancer / Iliad 78`);
       fs.writeFileSync(portable, content, { encoding: 'utf-8' });
     });
   });