+++ src/ts/typings/stancer.ts
@@ -15,7 +15,5 @@ export interface StancerSettings {
   };
   page_type: string;
   title: string;
-  supports: string[]
+  supports: string[];
 }
-
-